Cleanup introspection code
